[refactor] 코드리뷰 반영해서 리팩토링 진행

+++ src/main/java/com/moyobab/server/participant/dto/ParticipantResponseDto.java
@@ -11,6 +11,6 @@ public class ParticipantResponseDto {
     private Long id;
     private Long userId;
     private Long groupOrderId;
-    private int orderAmount;
+    private Long orderAmount;
     private boolean paid;
 }
